+++ src/Kappa.cpp
@@ -38,10 +38,10 @@ void Kappa::set_kappa(double k) { kappa = k; }
 void Kappa::set_p(double ucom, double bfield, double betaeff, double r, double fsc) {
     pmax = std::max(max_p(ucom, bfield, betaeff, r, fsc), pmax);
 
-    double pinc = (log10(pmax) - log10(pmin)) / (p.size() - 1);
+    double pinc = (log10(pmax) - log10(pmin)) / static_cast<double>(p.size() - 1);
 
     for (size_t i = 0; i < p.size(); i++) {
-        p[i] = pow(10., log10(pmin) + i * pinc);
+        p[i] = pow(10., log10(pmin) + static_cast<double>(i) * pinc);
         gamma[i] = pow(pow(p[i] / (mass_gr * constants::cee), 2.) + 1., 1. / 2.);
     }
 }
@@ -50,10 +50,10 @@ void Kappa::set_p(double ucom, double bfield, double betaeff, double r, double f
 void Kappa::set_p(double gmax) {
     pmax = pow(pow(gmax, 2.) - 1., 1. / 2.) * mass_gr * constants::cee;
 
-    double pinc = (log10(pmax) - log10(pmin)) / (p.size() - 1);
+    double pinc = (log10(pmax) - log10(pmin)) / static_cast<double>(p.size() - 1);
 
     for (size_t i = 0; i < p.size(); i++) {
-        p[i] = pow(10., log10(pmin) + i * pinc);
+        p[i] = pow(10., log10(pmin) + static_cast<double>(i) * pinc);
         gamma[i] = pow(pow(p[i] / (mass_gr * constants::cee), 2.) + 1., 1. / 2.);
     }
 }
