+++ src/Thermal.cpp
@@ -27,10 +27,10 @@ void Thermal::set_p() {
 
     pmin = pow(pow(gmin, 2.) - 1., 1. / 2.) * mass_gr * constants::cee;
     pmax = pow(pow(gmax, 2.) - 1., 1. / 2.) * mass_gr * constants::cee;
-    pinc = (log10(pmax) - log10(pmin)) / (p.size() - 1);
+    pinc = (log10(pmax) - log10(pmin)) / static_cast<double>(p.size() - 1);
 
     for (size_t i = 0; i < p.size(); i++) {
-        p[i] = pow(10., log10(pmin) + i * pinc);
+        p[i] = pow(10., log10(pmin) + static_cast<double>(i) * pinc);
         gamma[i] = pow(pow(p[i] / (mass_gr * constants::cee), 2.) + 1., 1. / 2.);
     }
 }
